add ReversePrivateRoutes
This makes it such that once a user has logged in or signed up, they are only able to access their respective dashborad route only.
Any attempt at navigating to other routes will result to getting navigated back to their dashboard route.
This feature should be such that it checks their trading number and determines which dashboard to route them back to.
If for some wild reason the Trading Number does not match the criteria set to navigate a user to their respective dashboard, then navigate them to '/' route.